reorganize code hyerarchy
