+++ src/controller/mod.rs
@@ -1,0 +1,11 @@
+mod aqi;
+mod index;
+
+use router::Router;
+
+pub fn initialize() -> Router {
+  let mut router: Router = Router::new();
+  router.get("/", index::index, "index");
+  router.post("/aqi/", aqi::parse, "parse");
+  router
+}
