+++ src/value.rs
@@ -1,0 +1,5 @@
+#[derive(Debug)]
+pub struct Aqi {
+  pub value: i16,
+  pub time: ::chrono::NaiveDateTime,
+}
